Rewire capa_plugin configuration with main configuration object
